Move bootloader vz code to vf package
Still trying to group cmdline processing/generation together, and to
keep vz code separate, while still sharing as much as possible.
