+++ pkg/config/bootloader.go
@@ -2,17 +2,11 @@ package config
 
 import (
 	"fmt"
-	"os"
 
 	"github.com/crc-org/vfkit/pkg/util"
-
-	"github.com/Code-Hex/vz/v3"
-	"github.com/h2non/filetype"
-	"github.com/h2non/filetype/matchers"
 )
 
 type Bootloader interface {
-	ToVzBootloader() (vz.BootLoader, error)
 	FromOptions(options []option) error
 }
 
@@ -36,47 +30,6 @@ func NewLinuxBootloader(vmlinuzPath, kernelCmdLine, initrdPath string) *LinuxBoo
 	}
 }
 
-func isKernelUncompressed(filename string) (bool, error) {
-	file, err := os.Open(filename)
-	if err != nil {
-		return false, err
-	}
-	defer file.Close()
-
-	buf := make([]byte, 2048)
-	_, err = file.Read(buf)
-	if err != nil {
-		return false, err
-	}
-	kind, err := filetype.Match(buf)
-	if err != nil {
-		return false, err
-	}
-	// uncompressed ARM64 kernels are matched as a MS executable, which is
-	// also an archive, so we need to special case it
-	if kind == matchers.TypeExe {
-		return true, nil
-	}
-
-	return false, nil
-}
-
-func (bootloader *LinuxBootloader) ToVzBootloader() (vz.BootLoader, error) {
-	uncompressed, err := isKernelUncompressed(bootloader.VmlinuzPath)
-	if err != nil {
-		return nil, err
-	}
-	if !uncompressed {
-		return nil, fmt.Errorf("kernel must be uncompressed, %s is a compressed file", bootloader.VmlinuzPath)
-	}
-
-	return vz.NewLinuxBootLoader(
-		bootloader.VmlinuzPath,
-		vz.WithCommandLine(bootloader.KernelCmdLine),
-		vz.WithInitrd(bootloader.InitrdPath),
-	)
-}
-
 func (bootloader *LinuxBootloader) FromOptions(options []option) error {
 	for _, option := range options {
 		switch option.key {
@@ -98,24 +51,6 @@ func NewEFIBootloader(efiVariableStorePath string, createVariableStore bool) *EF
 		EFIVariableStorePath: efiVariableStorePath,
 		CreateVariableStore:  createVariableStore,
 	}
-}
-
-func (bootloader *EFIBootloader) ToVzBootloader() (vz.BootLoader, error) {
-	var efiVariableStore *vz.EFIVariableStore
-	var err error
-
-	if bootloader.CreateVariableStore {
-		efiVariableStore, err = vz.NewEFIVariableStore(bootloader.EFIVariableStorePath, vz.WithCreatingEFIVariableStore())
-	} else {
-		efiVariableStore, err = vz.NewEFIVariableStore(bootloader.EFIVariableStorePath)
-	}
-	if err != nil {
-		return nil, err
-	}
-
-	return vz.NewEFIBootLoader(
-		vz.WithEFIVariableStore(efiVariableStore),
-	)
 }
 
 func (bootloader *EFIBootloader) FromOptions(options []option) error {
